feat: abstracted widget
