+++ frappe/public/js/frappe/views/desk/base_widget.js
@@ -1,0 +1,7 @@
+export default class Widget {
+	constructor(opts) {
+		Object.assign(this, opts);
+		this.make_widget();
+		this.setup_events();
+	}
+}
